integrated link fixed, successful communication between ESP32 and GUI

+++ skeleton/ESP32_Vehicle.h
@@ -121,7 +121,7 @@ void initSpeedServo();
 void initCapacitorPin();
 
 void setSteeringAngle(float angle);
-void setServoSpeed(float duty);
+void setServoSpeed(uint32_t duty);
 float calculateSteeringAngle();
 float calculateServoSpeed();
